directory Bombermaaan/: - added Makefile (Linux Makefile from linux_port, rev. 521) - added Makefile.win (Mingw32 Makefile from linux_port, rev. 519)
note: in Linux you can type make to compile (in src/) dir. first the 
resource library will be generated in RESGEN/
It is the replacement for resource dlls in win32.
Of course all IMAGE and SOUND files mentioned in the resource.rc in the 
RES(32) dir need to be present in the IMAGE and SOUND dirs in RES(32).

If you want to use mingw32 you have to adjust the variables in the 
Makefile.win files.

A mingw32 cross-compiler can be used in Linux to generate Win32 
binaries. You can type CROSS=1 make -f Makefile.win in the Bombermaaan/ dir.
You will need the resource dlls which can be compiled with the same 
command in the RES(32) dir.

But as a normal person, you would rather prefer to build the binaries 
nativetly ;-)
to do so, install the SDL mingw32 sdk and build SDL_mixer with mingw32.
Then adjust the paths in the Makefile.win and type
mingw32-make.exe -f Makefile.win

if you have acheived to build an executable successfully make sure that 
you have some levels in the current dir.
In Linux you may also place them in your ~/.Bombermaaan directory (which 
will be created upon the first call of the executable). Check the log 
file as well: ~/.Bombermaaan/log.txt

- added winreplace.cpp, winreplace.h (necessary for Linux compilation);
taken from linux_port rev. 519.

- added CSDLInput.cpp, CSDLInput.h, CSDLVideo.cpp, CSDLVideo.h
taken from linux_port rev. 519 (CSDLVideo.h), rev. 512 (rest)
CDirectInput/CDirectDraw replacements for Linux

- added BombermaaaanIco.h
taken from linux_port rev. 519. This is a char array representation of
Bombermaaaan.ico as a bitmap. It is necessary for the SDL_WM_SetIcon 
call which sets the application's icon.


git-svn-id: https://svn.code.sf.net/p/bombermaaan/code@535 7fe5b47a-8640-0410-b185-a85574715510